made Norm2Min accept decimal since it may be passed results of calculations of size.
git-svn-id: http://tradelink.googlecode.com/svn/trunk@168 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ TradeLib/Box.cs
@@ -174,9 +174,9 @@ namespace TradeLib
         /// </summary>
         /// <param name="size">The size.</param>
         /// <returns></returns>
-        protected int Norm2Min(int size)
+        protected int Norm2Min(decimal size)
         {
-            int wmult = (int)Math.Ceiling((decimal)size/MINSIZE);
+            int wmult = (int)Math.Ceiling(size/MINSIZE);
             return wmult * MINSIZE;
         }
